refactor deprecated activites into /deprecated
