+++ app/src/androidTest/java/com/github/fribourgsdp/radio/FireBaseTestActivityTest.kt
@@ -1,13 +1,12 @@
 package com.github.fribourgsdp.radio
 
 import android.content.Context
-import android.content.Intent
-import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso
 import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.ext.junit.rules.ActivityScenarioRule
+import com.github.fribourgsdp.radio.deprecated.FireBaseTestActivity
 import org.hamcrest.Matchers
 import org.junit.After
 import org.junit.Before
