+++ app/src/main/AndroidManifest.xml
@@ -34,7 +34,7 @@
             android:launchMode= "singleInstance"
             android:exported="false" />
         <activity
-            android:name=".VoiceOverIPActivity"
+            android:name=".deprecated.VoiceOverIPActivity"
             android:exported="false" />
         <activity
             android:name=".data.view.PlaylistsFragmentHolderActivity"
@@ -72,7 +72,7 @@
             android:exported="false"
             android:launchMode="singleTop" />
         <activity
-            android:name=".FireBaseTestActivity"
+            android:name=".deprecated.FireBaseTestActivity"
             android:exported="false" />
         <activity
             android:name=".config.SettingsActivity"
