+++ app/src/main/java/com/github/fribourgsdp/radio/deprecated/FireBaseTestActivity.kt
@@ -1,7 +1,10 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.deprecated
 
 import android.content.Intent
 import android.os.Bundle
+import com.github.fribourgsdp.radio.FirestoreDatabase
+import com.github.fribourgsdp.radio.MainActivity
+import com.github.fribourgsdp.radio.R
 import com.github.fribourgsdp.radio.config.MyAppCompatActivity
 import com.github.fribourgsdp.radio.data.Genre
 import com.github.fribourgsdp.radio.data.Playlist
