+++ app/src/main/res/layout/activity_voice_over_ip.xml
@@ -6,7 +6,7 @@
     android:id="@+id/activity_main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".VoiceOverIPActivity">
+    tools:context=".deprecated.VoiceOverIPActivity">
 
 
     <TextView
